didi_poi/requirements.txt
flask==2.3.0
werkzeug==2.3.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.0
Pillow==10.0.0

